Convert data list to tuple at array append.
Sql Server requires a tuple or dictionary for insert.
Rather than converting the line straight to a tuple and lose functionality like list.pop(), conversion is now happening at the very last time that the row values are touched (when appending it to the input_array.

Bug fix for bug #35

Signed-off-by: Gerald Venzl <gerald.venzl@gmail.com>